+++ example/redis_bench_shuttle.ml
@@ -1,19 +1,18 @@
 open! Core
 open! Async
-open Shuttle
+open Shuttle.Std
 
 let unlink f = Deferred.ignore_m (Monitor.try_with (fun () -> Unix.unlink f))
 
 let run sock =
   let%bind () = unlink sock in
   let%bind server =
-    Tcp.Server.create_sock
+    Shuttle.Connection.listen
+      ~input_buffer_size:0x1000
+      ~output_buffer_size:0x1000
       ~on_handler_error:`Raise
       (Tcp.Where_to_listen.of_file sock)
-      (fun _addr sock ->
-      let fd = Socket.fd sock in
-      let reader = Input_channel.create ~buf_len:0x1000 fd in
-      let writer = Output_channel.create ~buf_len:0x1000 fd in
+      (fun _addr reader writer ->
       let pipe_r = Input_channel.pipe reader in
       Pipe.iter pipe_r ~f:(fun buf ->
         String.iter buf ~f:(fun ch ->