+++ http/test/helper.ml
@@ -1,7 +1,5 @@
 open! Core
 open! Async
-open! Shuttle
-open! Shuttle_http
 
 let rec connect port =
   match%bind
@@ -26,15 +24,17 @@ let with_client port ~f =
 ;;
 
 let with_server ?error_handler ?read_header_timeout handler ~f =
+  let open Shuttle_http in
   let%bind server =
-    Tcp.Server.create_sock
+    Shuttle.Connection.listen
+      ~input_buffer_size:0x4000
+      ~output_buffer_size:0x4000
       ~max_accepts_per_batch:64
       ~on_handler_error:`Raise
       Tcp.Where_to_listen.of_port_chosen_by_os
-      (fun _addr sock ->
-      let fd = Socket.fd sock in
+      (fun _addr reader writer ->
       let server =
-        Shuttle_http.Server.create ?read_header_timeout ?error_handler ~buf_len:0x4000 fd
+        Shuttle_http.Server.create ?read_header_timeout ?error_handler reader writer
       in
       Server.run server (handler server))
   in