+++ shuttle/lib/shuttle.ml
@@ -1,3 +1,10 @@
 module Input_channel = Input_channel
 module Output_channel = Output_channel
+module Connection = Connection
 module Bytebuffer = Bytebuffer
+
+module Std = struct
+  module Input_channel = Input_channel
+  module Output_channel = Output_channel
+  module Bytebuffer = Bytebuffer
+end