+++ CHANGES.md
@@ -3,7 +3,7 @@
 * Forward ssl connection details to shuttle_ssl server handlers
 * Remove listen and with_connection from shuttle_ssl. Use upgrade_server_connection and upgrade_client_connection with the regular Connection module provided within Shuttle.
 * Allow accessing the underlying buffer capacity for channels
-* Remove connection module from Shuttle
+* Add helper method that creates server connections using inet sockets
 
 # 0.8.1